+++ EventListener/SendConfirmationEmail.php
@@ -23,17 +23,14 @@
 
 namespace Payzen\EventListener;
 
+use Payzen\Model\PayzenConfigQuery;
 use Payzen\Payzen;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Thelia\Action\BaseAction;
 use Thelia\Core\Event\Order\OrderEvent;
 use Thelia\Core\Event\TheliaEvents;
-use Thelia\Core\Template\ParserInterface;
-use Thelia\Log\Tlog;
 use Thelia\Mailer\MailerFactory;
-use Thelia\Model\ConfigQuery;
-use Thelia\Model\Lang;
-use Thelia\Model\MessageQuery;
 
 /**
  * Payzen payment module
@@ -42,18 +39,11 @@ use Thelia\Model\MessageQuery;
  */
 class SendConfirmationEmail extends BaseAction implements EventSubscriberInterface
 {
-    /**
-     * @var MailerFactory
-     */
+    /** @var MailerFactory */
     protected $mailer;
-    /**
-     * @var ParserInterface
-     */
-    protected $parser;
 
-    public function __construct(ParserInterface $parser, MailerFactory $mailer)
+    public function __construct(MailerFactory $mailer)
     {
-        $this->parser = $parser;
         $this->mailer = $mailer;
     }
 
@@ -66,62 +56,59 @@ class SendConfirmationEmail extends BaseAction implements EventSubscriberInterfa
     }
 
     /**
-     * Checks if we are the payment module for the order, and if the order is paid,
-     * then send a confirmation email to the customer.
+     * @param OrderEvent $event
      *
-     * @params OrderEvent $order
+     * @throws \Exception if the message cannot be loaded.
      */
-    public function update_status(OrderEvent $event)
+    public function sendConfirmationEmail(OrderEvent $event)
     {
-        $payzen = new Payzen();
+        if (PayzenConfigQuery::read('send_confirmation_message_only_if_paid')) {
+            // We send the order confirmation email only if the order is paid
+            $order = $event->getOrder();
 
-        if ($event->getOrder()->isPaid() && $payzen->isPaymentModuleFor($event->getOrder())) {
-            $contact_email = ConfigQuery::read('store_email', false);
-            $lang = Lang::getDefaultLanguage();
-            $locale = $lang->getLocale();
-
-            Tlog::getInstance()->debug("Sending confirmation email from store contact e-mail $contact_email");
-
-            if ($contact_email) {
-                $message = MessageQuery::create()
-                    ->filterByName(Payzen::CONFIRMATION_MESSAGE_NAME)
-                    ->findOne();
-
-                if (false === $message) {
-                    throw new \Exception(sprintf("Failed to load message '%s'.", Payzen::CONFIRMATION_MESSAGE_NAME));
-                }
-
-                $order = $event->getOrder();
-                $customer = $order->getCustomer();
-
-                $this->parser->assign('order_id', $order->getId());
-                $this->parser->assign('order_ref', $order->getRef());
-                $this->parser->assign('locale', $locale);
-
-                $message
-                    ->setLocale($order->getLang()->getLocale());
-
-                $instance = \Swift_Message::newInstance()
-                    ->addTo($customer->getEmail(), $customer->getFirstname()." ".$customer->getLastname())
-                    ->addFrom($contact_email, ConfigQuery::read('store_name'))
-                ;
-
-                // Build subject and body
-                $message->buildMessage($this->parser, $instance);
-
-                $this->getMailer()->send($instance);
-
-                Tlog::getInstance()->debug("Confirmation email sent to customer ".$customer->getEmail());
+            if (! $order->isPaid() && $order->getPaymentModuleId() == Payzen::getModuleId()) {
+                $event->stopPropagation();
             }
-        } else {
-            Tlog::getInstance()->debug("No confirmation email sent (order not paid, or not the proper payment module).");
+        }
+    }
+
+    /**
+     * Checks if order payment module is paypal and if order new status is paid, send an email to the customer.
+     *
+     * @param OrderEvent $event
+     * @param $eventName
+     * @param EventDispatcherInterface $dispatcher
+     *
+     * @throws \Propel\Runtime\Exception\PropelException
+     */
+    public function updateStatus(OrderEvent $event, $eventName, EventDispatcherInterface $dispatcher)
+    {
+        $order = $event->getOrder();
+
+        if ($order->isPaid() && $order->getPaymentModuleId() === Payzen::getModuleId()) {
+            if (PayzenConfigQuery::read('send_payment_confirmation_message')) {
+                $this->mailer->sendEmailToCustomer(
+                    Payzen::CONFIRMATION_MESSAGE_NAME,
+                    $order->getCustomer(),
+                    [
+                        'order_id'  => $order->getId(),
+                        'order_ref' => $order->getRef()
+                    ]
+                );
+            }
+
+            // Send confirmation email if required.
+            if (Payzen::getConfigValue('send_confirmation_message_only_if_paid')) {
+                $dispatcher->dispatch(TheliaEvents::ORDER_SEND_CONFIRMATION_EMAIL, $event);
+            }
         }
     }
 
     public static function getSubscribedEvents()
     {
         return array(
-            TheliaEvents::ORDER_UPDATE_STATUS => array("update_status", 128)
+            TheliaEvents::ORDER_UPDATE_STATUS           => array("updateStatus", 128),
+            TheliaEvents::ORDER_SEND_CONFIRMATION_EMAIL => array("sendConfirmationEmail", 129)
         );
     }
 }